+++ docs/docs/.vuepress/client.js
@@ -6,6 +6,10 @@ import SceneClearColor from './components/SceneClearColor.vue'
 import SkyBox from './components/SkyBox.vue'
 import GLTFLoaderComponent from './components/GLTFLoaderComponent.vue'
 import GLTFLoaderFunction from './components/GLTFLoaderFunction.vue'
+import FBXLoaderComponent from './components/FBXLoaderComponent.vue'
+import FBXLoaderFunction from './components/FBXLoaderFunction.vue'
+import OBJLoaderComponent from './components/OBJLoaderComponent.vue'
+import OBJLoaderFunction from './components/OBJLoaderFunction.vue'
 import Popup from './components/Popup.vue'
 import MovableElement from './components/MovableElement.vue'
 
@@ -18,6 +22,10 @@ export default defineClientConfig({
     app.component('SkyBox', SkyBox)
     app.component('GLTFLoaderComponent', GLTFLoaderComponent)
     app.component('GLTFLoaderFunction', GLTFLoaderFunction)
+    app.component('FBXLoaderComponent', FBXLoaderComponent)
+    app.component('FBXLoaderFunction', FBXLoaderFunction)
+    app.component('OBJLoaderComponent', OBJLoaderComponent)
+    app.component('OBJLoaderFunction', OBJLoaderFunction)
     app.component('Popup', Popup)
     app.component('MovableElement', MovableElement)
   }