+++ src/utils/Renderer.ts
@@ -1,9 +1,9 @@
 import * as THREE from 'three'
 
 export default function (): THREE.WebGLRenderer {
-  const Renderer = new THREE.WebGLRenderer({
+  const renderer = new THREE.WebGLRenderer({
     antialias: true,
     alpha: true
   })
-  return Renderer
+  return renderer
 }